+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - [DemoApp] Bug #107 Fixed : Colour pop-up displays wrong color
 - [DemoApp] Bug #102 Fixed : Card bad display
 - [DemoApp] Bug #101 Fixed : Progress indicator not centered
+- [DemoApp] Bug #100 Fixed : Update thumb on Components screen
 
 ## [0.2.0] - 2022-05-05
 ### Changed