+++ CHANGELOG.md
@@ -7,9 +7,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+### Bugs
 - [DemoApp] Bug #107 Fixed : Colour pop-up displays wrong color
 - [DemoApp] Bug #102 Fixed : Card bad display
 - [DemoApp] Bug #101 Fixed : Progress indicator not centered
+
+### Changed
+- [DemoApp] #112 Remove Imagery and Iconography from GuideLines screen
 
 ## [0.2.0] - 2022-05-05
 ### Changed